rename integration-tests to tests
